Fix volsmile CLI defaults and entrypoint naming
